refactor(dashboard): extract the project controllers section
Third composition-section extraction (B1): the ten project controllers, the
two resolvers, and the shared project surface refresh move into
src/dashboard/sections/projectControllers.ts. Late-bound collaborators
arrive as getters; the source-text anchors in the safety and webview checks
follow the constructions into the section files. No behavior change.
